Tradition Finisher Change
Changed the Tradition Finisher Social Policy from 4 free food buildings
to +3 food per city.  Retained +15% Growth per city policy in the
finisher.

Also added text to the Leaders.sql file for this change.

It hasn't been tested to confirm that it works, but it should.  You may
want to alter the +3 food per city to something different.  It's an
arbitrary value.