Jesuit Education Bug
The bug is in Firaxis code.  Workaround by changing the belief to a
different effect until Firaxis code is patched.  Changed Jesuit
Education from allowing the purchase of Science buildings to giving
science in each city with a specialist.  (Same as what GuruShip did).
Changed GuruShip back to the vanilla code giving net production in
cities with a specialist.

+4 science with Jesuit Education.
+6 production with GuruShip.

Since Science is a better, more powerful yield than production.